requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
Django==2.2.16
django-autocomplete-light==3.5.1
django-bootstrap-static==4.2.1
django-bootstrap4==1.1.1
django-cleanup==4.0.0
django-generic-m2m==0.3.1
django-markdownx==3.0.1
django-ordered-model==3.4.1
django-responsive-images==1.2.0
django-static-fontawesome==5.10.2
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
Pillow==6.1.0
pyembed==1.3.3
pyembed-markdown==1.1.0
pymdown-extensions==8.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
zipp==1.2.0
